Reimplement the output pacing algorithm.  The output pacing should be a function of the size of each command sent to the irc server. Hugo Haas suggested adding int(s/100) seconds delay to the next output where s is the total size in bytes of the command.  Liam Quin says that the server flood algorithms are typically also dependent on the total number of bytes a client has inserted in a channel output queue awaiting processing.  Thus, a slow client can cause a fast client to be disconnected due to flooding prevention.
The new algorithm attempts to adjust for the total number of bytes waiting
to be sent to the server.  It also handles the case were multiple channels
each caused new output to be queued; the old algorithm tried to respond
immediately to each channel, which might have been triggering flood control.